LU-610 fix the wrong data type in find_value_cmp
in function find_value_cmp(lustre/utils/liblustreapi.c), the data type for
parameters "file" and "limit" is converted from "unsigned long long" type
into "unsigned int" type by mistake, and the file size should be defined
as "unsigned long long" type too.

Reviewed-on: http://review.whamcloud.com/1401
Tested-by: Hudson
Tested-by: Maloo <whamcloud.maloo@gmail.com>

+++ lustre/utils/liblustreapi.c
@@ -1708,8 +1708,9 @@ int llapi_file_lookup(int dirfd, const char *name)
  * --------------------------------------
  * Note: 5th actually means that the value is within the interval
  * (limit - margin, limit]. */
-static int find_value_cmp(unsigned int file, unsigned int limit, int sign,
-                          int negopt, unsigned long long margin, int mds)
+static int find_value_cmp(unsigned long long file, unsigned long long limit,
+                          int sign, int negopt, unsigned long long margin,
+                          int mds)
 {
         int ret = -1;
